KAN-48 fix: change sidebar, deleted SidebarMenu, add YesAndNoModal
