+++ src/app/profile/layout.tsx
@@ -1,10 +1,10 @@
-import { SidebarMenu } from '@/widgets/Sidebar'
+import { Sidebar } from '@/widgets/Sidebar'
 import { ReactNode } from 'react'
 
 export default function ProfileLayout({ children }: { children: ReactNode }) {
    return (
       <div className="mx-auto flex w-full max-w-[1280px] flex-col items-start px-[60px]">
-         <SidebarMenu />
+         <Sidebar />
          <div className="border-dark-300 ml-[180px] min-h-screen border-l">{children}</div>
       </div>
    )
