+++ src/entities/common/ui/index.ts
@@ -1,0 +1,1 @@
+export { YesAndNoModal } from './YesAndNoModal'
