+++ src/shared/hook/useAppDispatch.ts
@@ -1,0 +1,4 @@
+import { useDispatch } from 'react-redux'
+import { AppDispatch } from '../store'
+
+export const useAppDispatch = () => useDispatch<AppDispatch>()
