+++ src/widgets/Sidebar/index.ts
@@ -1,5 +1,3 @@
 export * from './Sidebar'
 export { sidebarData } from './sidebarData'
 export type { SidebarItemType } from './sidebarData'
-export * from './LogoutModal'
-export * from './SidebarMenu'
